+++ mrz-3-edition/chapters/2/exercises/1/maxMinFinder.m
@@ -15,26 +15,20 @@
 %}
 
 
-function [c] = fromBase10FractionalPartConversion (a, b, m)
-% FROMBASE10FRACTIONALPARTCONVERSION Convert a fractional part of a number
-%     from base 10 to number c in base b (with at most m decimal places)
+function [firstGreaterThan, firstLowerThan] = minMaxFinder (aMin, aMax)
+% MINMAXFINDER Finds
+% - the first number greater than the min value input
+% - the first number lower than the max value input
 %
-% [c] = fromBase10FractionalPartConversion (a, b, m)
+% [firstGreaterThan, firstLowerThan] = minMaxFinder (aMin, aMax)
 %
 % Input:
-% a - fractional part of number to convert from base 10 to base b
-% b - base form which convert number a; b must >= 2
-% m - max number of precision digits in result
+% aMin - the min value
+% aMax - the max value
 %
 % Output:
-% c - fractional part of number converted from base 10 to base b
-%     (array-like)
+% firstGreaterThan - the first number greater than the min value input
+% firstLowerThan- the first number lower than the max value input
 
-current_fractional_part = a;  % decimal part used in multplications
-c = [];  % result array-like
-for decimal_places_current_precision = 0 : m
-    current_fractional_part = current_fractional_part * b;
-    int_part = floor(current_fractional_part);
-    c = [c int_part];  % add just computed decimal digit
-    current_fractional_part = current_fractional_part - int_part;  % get only fractional part
-end
+firstGreaterThan = aMin + eps * 3;
+firstLowerThan = aMax - eps * 3;